[moyora] RoomRepository에 RoomVo를 찾아 목적지 설정하는 서비스 완성

+++ src/main/java/com/project/side/moyora/service/RoomRepositoryService.java
@@ -5,6 +5,6 @@ import com.project.side.moyora.vo.RoomVo;
 
 public interface RoomRepositoryService {
     RoomVo createRoom(GuestVo guest);
-    void deleteRoom(int roomNo);
-    RoomVo findRoom(int roomNo);
+    void deleteRoom(long roomNo);
+    RoomVo findRoom(long roomNo);
 }
